Vendor linux-pe and reorganize deps
Dropped the VTIL-Core and linux-pe submodules - the latter is now
vendored directly into external/linux-pe so we can patch it if needed.
Added msvc-wine for cross-compiling from Linux.

Moved the old NoVmp/ tree to legacy/; it's not the active code path
anymore but worth keeping for reference. The new code lives in lib/
and src/ on a proper CMake/C++23 build.

Also pulled in sample binaries (both clean and VMProtected versions)
for testing the lifter against known outputs.

Couple of small fixes in the core: streamlined the disasm stream
interface, cleaned up some LLVM generation in vmp_to_llvm, and
fixed how we resolve image descriptors during entry discovery.

+++ lib/vmp/image_desc.hpp
@@ -13,7 +13,7 @@ namespace vmp {
 // Result structure for lifted routine
 struct LiftedRoutine {
     std::unique_ptr<ir::LiftContext> context;
-    uint32_t handler_count = 0;
+    size_t handler_count = 0;
     size_t instructions_before = 0;
     size_t instructions_after = 0;
 };
